refactor: move resolve approvers to step
